Develop

+++ tests/DataStorage/Cache/Connection/MemCachedTest.php
@@ -2,7 +2,7 @@
 /**
  * Jingga
  *
- * PHP Version 8.1
+ * PHP Version 8.2
  *
  * @package   tests
  * @copyright Dennis Eichhorn
@@ -20,10 +20,10 @@ use phpOMS\DataStorage\Cache\Connection\MemCached;
 use phpOMS\Utils\TestUtils;
 
 /**
- * @testdox phpOMS\tests\DataStorage\Cache\Connection\MemCachedTest: Memcache connection
- *
  * @internal
  */
+#[\PHPUnit\Framework\Attributes\CoversClass(\phpOMS\DataStorage\Cache\Connection\MemCached::class)]
+#[\PHPUnit\Framework\Attributes\TestDox('phpOMS\tests\DataStorage\Cache\Connection\MemCachedTest: Memcache connection')]
 final class MemCachedTest extends \PHPUnit\Framework\TestCase
 {
     protected MemCached $cache;
@@ -44,14 +44,15 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
 
     protected function tearDown() : void
     {
+        if (!isset($this->cache)) {
+            return;
+        }
+
         $this->cache->flushAll();
     }
 
-    /**
-     * @testdox The memcached connection has the expected default values after initialization
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('The memcached connection has the expected default values after initialization')]
     public function testDefault() : void
     {
         self::assertEquals(CacheType::MEMCACHED, $this->cache->getType());
@@ -71,11 +72,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         );
     }
 
-    /**
-     * @testdox The connection to a cache can be established (none-existing directories get created)
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('The connection to a cache can be established (none-existing directories get created)')]
     public function testConnect() : void
     {
         $cache = new MemCached($GLOBALS['CONFIG']['cache']['memcached']);
@@ -85,11 +83,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         self::assertEquals((int) $GLOBALS['CONFIG']['cache']['memcached']['port'], $cache->getPort());
     }
 
-    /**
-     * @testdox Different cache data (types) can be set and returned
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Different cache data (types) can be set and returned')]
     public function testSetInputOutput() : void
     {
         $this->cache->set('key1', 'testVal');
@@ -120,11 +115,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         */
     }
 
-    /**
-     * @testdox Cache data can bet added and returned
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Cache data can bet added and returned')]
     public function testAddInputOutput() : void
     {
         self::assertTrue($this->cache->add('addKey', 'testValAdd'));
@@ -217,11 +209,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         self::assertFalse($this->cache->updateExpire('invalid', 2));
     }
 
-    /**
-     * @testdox Cache data cannot be added if it already exists
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Cache data cannot be added if it already exists')]
     public function testInvalidOverwrite() : void
     {
         self::assertTrue($this->cache->add('addKey', 'testValAdd'));
@@ -229,11 +218,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         self::assertEquals('testValAdd', $this->cache->get('addKey'));
     }
 
-    /**
-     * @testdox Existing cache data can be replaced
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Existing cache data can be replaced')]
     public function testReplace() : void
     {
         $this->cache->set('key4', 4);
@@ -243,21 +229,15 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         self::assertEquals(5, $this->cache->get('key4'));
     }
 
-    /**
-     * @testdox None-existing cache data cannot be replaced
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('None-existing cache data cannot be replaced')]
     public function testInvalidReplace() : void
     {
         self::assertFalse($this->cache->replace('keyInvalid', 5));
     }
 
-    /**
-     * @testdox Existing cache data can be deleted
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Existing cache data can be deleted')]
     public function testDelete() : void
     {
         $this->cache->set('key4', 4);
@@ -278,11 +258,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         self::assertTrue($this->cache->delete('invalid'));
     }
 
-    /**
-     * @testdox The cache correctly handles general cache information
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('The cache correctly handles general cache information')]
     public function testStats() : void
     {
         $this->cache->set('key1', 'testVal');
@@ -305,11 +282,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         */
     }
 
-    /**
-     * @testdox The cache can be flushed
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('The cache can be flushed')]
     public function testFlush() : void
     {
         $this->cache->set('key1', 'testVal');
@@ -326,22 +300,16 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         self::assertGreaterThanOrEqual(0, $this->cache->stats()['size']);
     }
 
-    /**
-     * @testdox Cache data can be set and returned with expiration limits
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Cache data can be set and returned with expiration limits')]
     public function testUnexpiredInputOutput() : void
     {
         $this->cache->set('key1', 'testVal', 1);
         self::assertEquals('testVal', $this->cache->get('key1'));
     }
 
-    /**
-     * @testdox Expired cache data cannot be returned
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Expired cache data cannot be returned')]
     public function testExpiredInputOutput() : void
     {
         $this->cache->set('key2', 'testVal2', 1);
@@ -350,11 +318,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         self::assertNull($this->cache->get('key2', 1));
     }
 
-    /**
-     * @testdox Cache data can be flushed by expiration date
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Cache data can be flushed by expiration date')]
     public function testFlushExpired() : void
     {
         $this->cache->set('key6', 'testVal6', 1);
@@ -364,11 +329,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         self::assertNull($this->cache->get('key6', 0));
     }
 
-    /**
-     * @testdox A bad cache status will prevent all cache actions
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('A bad cache status will prevent all cache actions')]
     public function testBadCacheStatus() : void
     {
         TestUtils::setMember($this->cache, 'status', CacheStatus::FAILURE);
@@ -383,11 +345,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         self::assertEquals([], $this->cache->stats());
     }
 
-    /**
-     * @testdox Adding a invalid data type will throw an InvalidArgumentException
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Adding a invalid data type will throw an InvalidArgumentException')]
     public function testInvalidDataTypeAdd() : void
     {
         $this->expectException(\InvalidArgumentException::class);
@@ -395,11 +354,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         $this->cache->add('invalid', $this->cache);
     }
 
-    /**
-     * @testdox Setting a invalid data type will throw an InvalidArgumentException
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('Setting a invalid data type will throw an InvalidArgumentException')]
     public function testInvalidDataTypeSet() : void
     {
         $this->expectException(\InvalidArgumentException::class);
@@ -407,11 +363,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         $this->cache->set('invalid', $this->cache);
     }
 
-    /**
-     * @testdox A invalid host throws a InvalidConnectionConfigException
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('A invalid host throws a InvalidConnectionConfigException')]
     public function testInvalidCacheHost() : void
     {
         $this->expectException(\phpOMS\DataStorage\Cache\Exception\InvalidConnectionConfigException::class);
@@ -422,11 +375,8 @@ final class MemCachedTest extends \PHPUnit\Framework\TestCase
         $cache = new MemCached($db);
     }
 
-    /**
-     * @testdox A invalid port throws a InvalidConnectionConfigException
-     * @covers phpOMS\DataStorage\Cache\Connection\MemCached<extended>
-     * @group framework
-     */
+    #[\PHPUnit\Framework\Attributes\Group('framework')]
+    #[\PHPUnit\Framework\Attributes\TestDox('A invalid port throws a InvalidConnectionConfigException')]
     public function testInvalidCachePort() : void
     {
         $this->expectException(\phpOMS\DataStorage\Cache\Exception\InvalidConnectionConfigException::class);
